Removed old parser
